plugins/fa: fix custom symbol pane for Persian language
The symbol pane used to be refered by the path from the keyboard's data
dir, because that's where the language plugins were installed. But since
the restructure in b7884136fc4e ("Switch the plugin installation path to
Lomiri"), that's no longer the case.

Call `Qt.resolvedUrl()` on the layout file itself instead to solve the
path relative to the layout file, and fix how the symbol layout file
refers to the internal of the keyboard.