Reduce the weight of Bengali plugins
First, remove the word prediction database for bn-probhat. As it turns
out, lomiri-keyboard will automatically use the database for the
no-suffix variant automatically, so that saves up 1 copy of database.
This means the Probhat layout will have to depend on Jatiya
(traditional) layout, but since English's Dvorak layout do it too, I
guess it's fine.

And since Probhat have to depends on Jatiya anyway, also replace the
duplicated copy of symbol layout with the symlink. It's not much saving,
but why not?

And finally, reduce the size of the word completion database itself by
removing all but 1 of the source e-book. This reduces the size of the
database down from ~62M to ~19M. It's still the largest database, but at
least it's more resonable.

All these improvements should reduce the size of the rootfs to be under
certain device's limit again.